+++ code/__HELPERS/icons.dm
@@ -1098,6 +1098,14 @@ GLOBAL_LIST_INIT(freon_color_matrix, list("#2E5E69", "#60A2A8", "#A1AFB1", rgb(0
 /// Save file used in icon2base64. Used for converting icons to base64.
 GLOBAL_DATUM_INIT(dummySave, /savefile, new("tmp/dummySave.sav")) //Cache of icons for the browser output
 
+
+/// Generate a filename for this asset
+/// The same asset will always lead to the same asset name
+/// (Generated names do not include file extention.)
+/proc/generate_asset_name(file)
+	return "asset.[md5(fcopy_rsc(file))]"
+
+
 /**
   * Converts an icon to base64. Operates by putting the icon in the iconCache savefile,
   * exporting it as text, and then parsing the base64 from that.
@@ -1133,10 +1141,10 @@ GLOBAL_DATUM_INIT(dummySave, /savefile, new("tmp/dummySave.sav")) //Cache of ico
 		if (isfile(thing)) //special snowflake
 			var/name = sanitize_filename("[generate_asset_name(thing)].png")
 			if (!SSassets.cache[name])
-				register_asset(name, thing)
+				SSassets.transport.register_asset(name, thing)
 			for (var/thing2 in targets)
-				send_asset(thing2, key)
-			return "<img class='icon icon-misc' src=\"[url_encode(name)]\">"
+				SSassets.transport.send_assets(thing2, name)
+			return "<img class='icon icon-misc' src='[SSassets.transport.get_asset_url(name)]'>"
 		var/atom/A = thing
 		if (isnull(dir))
 			dir = A.dir
@@ -1158,11 +1166,11 @@ GLOBAL_DATUM_INIT(dummySave, /savefile, new("tmp/dummySave.sav")) //Cache of ico
 
 	key = "[generate_asset_name(I)].png"
 	if(!SSassets.cache[key])
-		register_asset(key, I)
+		SSassets.transport.register_asset(key, I)
 	for (var/thing2 in targets)
-		send_asset(thing2, key)
+		SSassets.transport.send_assets(thing2, key)
 
-	return "<img class='icon icon-[icon_state]' src=\"[url_encode(key)]\">"
+	return "<img class='icon icon-[icon_state]' src='[SSassets.transport.get_asset_url(key)]'>"
 
 /proc/icon2base64html(thing)
 	if (!thing)
