+++ code/game/machinery/embedded_controller/embedded_controller_base.dm
@@ -29,7 +29,6 @@
 	. = ..()
 	user.set_machine(src)
 	var/datum/browser/popup = new(user, "computer", name) // Set up the popup browser window
-	popup.set_title_image(user.browse_rsc_icon(src.icon, src.icon_state))
 	popup.set_content(return_text())
 	popup.open()
