+++ code/modules/admin/admin_verbs.dm
@@ -150,7 +150,8 @@ GLOBAL_PROTECT(admin_verbs_server)
 	/client/proc/toggle_hub,
 	/client/proc/mentor_memo, // YOGS - something stupid about "Mentor memos"
 	///client/proc/dump_memory_usage,
-	/client/proc/release_queue // Yogs -- Adds some queue-manipulation verbs
+	/client/proc/release_queue, // Yogs -- Adds some queue-manipulation verbs
+	/client/proc/toggle_cdn
 	)
 GLOBAL_LIST_INIT(admin_verbs_debug, world.AVerbsDebug())
 GLOBAL_PROTECT(admin_verbs_debug)
@@ -191,7 +192,8 @@ GLOBAL_PROTECT(admin_verbs_debug)
 	/client/proc/cmd_display_overlay_log,
 	/client/proc/reload_configuration,
 	/datum/admins/proc/create_or_modify_area,
-	/client/proc/debug_typeof // Yogs -- Adds a debug verb for getting the subtypes of something
+	/client/proc/debug_typeof, // Yogs -- Adds a debug verb for getting the subtypes of something
+	/client/proc/toggle_cdn
 	)
 GLOBAL_LIST_INIT(admin_verbs_possess, list(/proc/possess, /proc/release))
 GLOBAL_PROTECT(admin_verbs_possess)
