+++ code/modules/admin/verbs/beakerpanel.dm
@@ -65,7 +65,9 @@
 	set name = "Spawn reagent container"
 	if(!check_rights())
 		return
-
+	var/datum/asset/asset_datum = get_asset_datum(/datum/asset/simple/namespaced/common)
+	asset_datum.send()
+	//Could somebody tell me why this isn't using the browser datum, given that it copypastes all of browser datum's html
 	var/dat = {"
 		<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 		<html>
@@ -73,7 +75,7 @@
 				<meta http-equiv="Content-Type" content="text/html">
 				<meta http-equiv="X-UA-Compatible" content="IE=edge">
 				<meta charset='UTF-8'>
-				<link rel='stylesheet' type='text/css' href='common.css'>
+				<link rel='stylesheet' type='text/css' href='[SSassets.transport.get_asset_url("common.css")]'>
 				<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.js"></script>
 				<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.7/js/select2.full.min.js"></script>
 				<link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.7/css/select2.min.css">
