+++ code/modules/asset_cache/asset_cache_client.dm
@@ -12,10 +12,6 @@
 
 /// Process asset cache client topic calls for "asset_cache_preload_data=[HTML+JSON_STRING]
 /client/proc/asset_cache_preload_data(data)
-	/*var/jsonend = findtextEx(data, "{{{ENDJSONDATA}}}")
-	if (!jsonend)
-		CRASH("invalid asset_cache_preload_data, no jsonendmarker")*/
-	//var/json = html_decode(copytext(data, 1, jsonend))
 	var/json = data
 	var/list/preloaded_assets = json_decode(json)
 
@@ -26,19 +22,17 @@
 	sent_assets |= preloaded_assets
 
 
-/// Updates the client side stored html/json combo file used to keep track of what assets the client has between restarts/reconnects.
-/client/proc/asset_cache_update_json(verify = FALSE, list/new_assets = list())
+/// Updates the client side stored json file used to keep track of what assets the client has between restarts/reconnects.
+/client/proc/asset_cache_update_json()
 	if (world.time - connection_time < 10 SECONDS) //don't override the existing data file on a new connection
 		return
-	if (!islist(new_assets))
-		new_assets = list("[new_assets]" = md5(SSassets.cache[new_assets]))
 
-	src << browse(json_encode(new_assets|sent_assets), "file=asset_data.json&display=0")
+	src << browse(json_encode(sent_assets), "file=asset_data.json&display=0")
 
-/// Blocks until all currently sending browser assets have been sent.
+/// Blocks until all currently sending browse and browse_rsc assets have been sent.
 /// Due to byond limitations, this proc will sleep for 1 client round trip even if the client has no pending asset sends.
 /// This proc will return an untrue value if it had to return before confirming the send, such as timeout or the client going away.
-/client/proc/asset_flush(timeout = 50)
+/client/proc/browse_queue_flush(timeout = 50)
 	var/job = ++last_asset_job
 	var/t = 0
 	var/timeout_time = timeout
