+++ code/modules/asset_cache/asset_list_items.dm
@@ -104,7 +104,7 @@
 		/datum/asset/simple/irv
 	)
 
-/datum/asset/simple/changelog
+/datum/asset/simple/namespaced/changelog
 	assets = list(
 		"88x31.png" = 'html/88x31.png',
 		"bug-minus.png" = 'html/bug-minus.png',
@@ -126,37 +126,45 @@
 		"chrome-wrench.png" = 'html/chrome-wrench.png',
 		"changelog.css" = 'html/changelog.css'
 	)
+	parents = list("changelog.html" = 'html/changelog.html')
+	
 
 /datum/asset/group/goonchat
 	children = list(
 		/datum/asset/simple/jquery,
-		/datum/asset/simple/goonchat,
+		/datum/asset/simple/namespaced/goonchat,
 		/datum/asset/spritesheet/goonchat,
-		/datum/asset/simple/fontawesome
+		/datum/asset/simple/namespaced/fontawesome
 	)
 
 /datum/asset/simple/jquery
+	legacy = TRUE
 	assets = list(
 		"jquery.min.js"            = 'code/modules/goonchat/browserassets/js/jquery.min.js',
 	)
 
-/datum/asset/simple/goonchat
+/datum/asset/simple/namespaced/goonchat
+	legacy = TRUE
 	assets = list(
 		"json2.min.js"             = 'code/modules/goonchat/browserassets/js/json2.min.js',
 		"browserOutput.js"         = 'code/modules/goonchat/browserassets/js/browserOutput.js',
 		"browserOutput.css"	       = 'code/modules/goonchat/browserassets/css/browserOutput.css',
 		"browserOutput_white.css"  = 'code/modules/goonchat/browserassets/css/browserOutput_white.css',
 	)
+	parents = list(
+		//this list intentionally left empty (parent namespaced assets can't be referred to by name, only by generated url, and goonchat isn't smart enough for that. yet)
+	)
 
-/datum/asset/simple/fontawesome
+/datum/asset/simple/namespaced/fontawesome
+	legacy = TRUE
 	assets = list(
 		"fa-regular-400.eot"  = 'html/font-awesome/webfonts/fa-regular-400.eot',
 		"fa-regular-400.woff" = 'html/font-awesome/webfonts/fa-regular-400.woff',
 		"fa-solid-900.eot"    = 'html/font-awesome/webfonts/fa-solid-900.eot',
 		"fa-solid-900.woff"   = 'html/font-awesome/webfonts/fa-solid-900.woff',
-		"font-awesome.css"    = 'html/font-awesome/css/all.min.css',
 		"v4shim.css"          = 'html/font-awesome/css/v4-shims.min.css'
 	)
+	parents = list("font-awesome.css"    = 'html/font-awesome/css/all.min.css')
 
 /datum/asset/spritesheet/goonchat
 	name = "chat"
@@ -176,9 +184,25 @@
 
 	..()
 
+/datum/asset/simple/lobby
+	assets = list(
+		"playeroptions.css" = 'html/browser/playeroptions.css'
+	)
+
+/datum/asset/simple/namespaced/common
+	assets = list("padlock.png"	= 'html/padlock.png')
+	parents = list("common.css" = 'html/browser/common.css', "common.js" = 'html/browser/common.js')
+
 /datum/asset/simple/permissions
 	assets = list(
-		"padlock.png"	= 'html/padlock.png'
+		"search.js" = 'html/admin/search.js',
+		"panels.css" = 'html/admin/panels.css'
+	)
+
+/datum/asset/group/permissions
+	children = list(
+		/datum/asset/simple/permissions,
+		/datum/asset/simple/namespaced/common
 	)
 
 /datum/asset/simple/notes
