+++ code/modules/mob/dead/new_player/poll.dm
@@ -292,8 +292,8 @@
 				<head>
 				<meta charset='UTF-8'>
 				<meta http-equiv="X-UA-Compatible" content="IE=edge" />
-				<script src="jquery.min.js"></script>
-				<script src="jquery-ui.custom-core-widgit-mouse-sortable-min.js"></script>
+				<script src="[SSassets.transport.get_asset_url("jquery.min.js")]"></script>
+				<script src="[SSassets.transport.get_asset_url("jquery-ui.custom-core-widgit-mouse-sortable-min.js")]"></script>
 				<style>
 					#sortable { list-style-type: none; margin: 0; padding: 2em; }
 					#sortable li { min-height: 1em; margin: 0px 1px 1px 1px; padding: 1px; border: 1px solid black; border-radius: 5px; background-color: white; cursor:move;}
