+++ code/modules/mob/dead/observer/orbit.dm
@@ -105,3 +105,8 @@
 	data["misc"] = misc
 	data["npcs"] = npcs
 	return data
+	
+/datum/orbit_menu/ui_assets()
+	. = ..() || list()
+	. += get_asset_datum(/datum/asset/simple/orbit)
+
