+++ config/config.txt
@@ -4,6 +4,7 @@ $include game_options.txt
 $include secret.txt
 $include dbconfig.txt
 $include antag_rep.txt
+$include resources.txt
 
 # You can use the @ character at the beginning of a config option to lock it from being edited in-game
 # Example usage:
